Improved on fix for laggy pushing bug

+++ Stonephonia/Entity.cs
@@ -10,7 +10,7 @@ namespace Stonephonia
         public Color mColour = Color.White;
         public float mVelocity;
         public float mSpeedModifier;
-        public int mMaxSpeed = 0;
+        public float mMaxSpeed = 0;
         public int mCollisionOffset = 0;
 
         public Rectangle mCollisionRect
